Utility functions for development environment
Add stop-server function under expenses.server namespace and create
dev/user.clj to be the starting point of the repl.

Add functions to start a brower-repl, start, stop and restart the
development application server under user namespace.

For more information about this workflow please refer to
http://thinkrelevance.com/blog/2013/06/04/clojure-workflow-reloaded

Of course this is only one piece of the full Stuart Sierra's workflow.